Steve L Fixed an unsuppressed output that I created in my previous checkin.
Added option to export Excel .tsv files oriented so that each row
is one image instead of one bin. This is now the default action.
Inserted question in Prompts to allow user to select between image
rows and bin rows. This is now question 12 of 13.

Moved code for writing Excel .tsv file to a function WriteHistToExcel
because nearly identical code appeared in two places.

Added error checking and error dialogs to WriteHistToExcel.


git-svn-id: https://svn.broadinstitute.org/CellProfiler/trunk/CellProfiler@759 1fc53939-2000-0410-845c-e8453a809027