Add test_15_01_profile_example_all, run_pipeline, and profile_pipeline to Test_Pipeline
run_pipeline is just a convenience method for running a pipeline from its filename.
profile_pipeline calls run_pipeline using the cProfile module.
test_15_01_profile_example_all profiles the ExampleAllModulesPipeline, which is assumed to be on the users machine. Issues a warning and aborts the test if not found.

Note, this test will be lengthy, the pipeline has a long runtime (~1 hour on this authors machine).

Move some of the testdata in test_Pipeline from being hardcoded to being in tests/data