Add versions based on git or SVN timestamp from logs.
Change version number to be the UTC timestamp of the most recent
change to the repository, either in git or SVN.  The timestamp is
extracted from the git log or SVN info, and represented as
YYYYMMDDHHMMSS in integer form.  Also added a dotted_version, to be
updated by hand when releases are made, in the form X.Y.Z, currently
set to 2.0.0.  Added a helper variable, title_string, which gives both
the integer revision as well as the git hash or SVN revision, in
string form.  Also added tests for well-formedness of the version
number of dotted_version.  Renamed submodule utilities.version and
removed utilities.get_revision.  Fixes gh-261.