Speed up startup of CellProfiler on disconnected systems.
This changes the method CP uses to determine if it's within the Broad
Institute to use socket.getfqdn() instead of
socket.gethostbyaddr(socket.gethostname()).  The latter can take more
than a minute to time out on systems without a network connection.  In
the exporttodatabase tests, it uses both versions,, since testing is
less time critical and test machines can be misconfigured in various
ways.