Rewrite pipeline validation to simplify caching.
While tracking down issue #253, simplified the code for validating
pipeline and modules in a separate thread.  Changed the validation
queue to a heap, to allow the displayed module to be validated with
higher priority than other modules.  Changed pipeline settings_hash()
to use settings' unicode_value, to ensure that
pipeline.settings_hash() == pipeline.copy().settings_hash().  I
believe this last piece was the only one necessary for fixing the bug,
but the rest is still useful.