Make cellprofiler on OSX 64-bit work in headless mode.
Modified ConvertObjectsToImage, DispalyDataOnImage, TrackObjects,
DefineGrid to remove dependencies on WX and wxAgg backend of
matplotlib (Agg backend has all the necessary pieces).

Pulled some tools out of cpfigure.py (which imports wx) into
cpfigure_tools.py (which does not rely on wx), to support the changes
above.

Modified a test for LoadData to work on both 32- and 64-bit Mac, and
cpnose.py to be smarter about running in headless mode.