Modified the compiler and linker command line arguments for better OS-specific behavior.
The -O3 compiler flag is invalid for the Microsoft Visual C compiler (Windows SDK versions 6.1-7.1). Though this flag is safely ignored by the compiler, it is better to be as specific as possible.
The Microsoft Visual C linker (same versions) does not automatically build the manifest file, even though the documentation suggests it does. Omitting this caused compilation to fail. Adding the "/MANIFEST" argument fixes this.