Shared state dictionary handling, Req/reply/reply/reply pattern, and some debugging.
Shared state between imagesets is handled via the run_state dictionary
in the modules, fetched via get_dict().  For multiprocessing, it's
necessary to communicate this to the jobserver and to serve it back to
jobs (either on demand or when they start).

The analysis manager generates the shared dictionary in prepare_run,
    and snapshots it when the first imageset completes.  Each new
    group fetches this snapshot from the manager at its beginning.  We
    make sure the first imageset finishes (and the dictionaries are
    fetched) before others are allowed to start.

Add please_reply kwarg to Req.reply() and Communicable.reply().
  This change allows Req/Reply to become Req/[Reply/Reply/]+Reply.

Also added a bit of extra debugging to Communicable.recv()