Make CellProfiler able to run on numpy trunk.
Numpy has introduced two changes, recently:
- default casting for in-place operations is same_kind instead of unsafe.
  This prevents bugs like adding (in place) a float to an int.
- Fancy indexing with boolean arrays require the same size for the two arrays.
  Previously, they were clipped to each other's extents.

The first change will probably not be permanent until numpy 2.0, but is set in trunk,
and led to the discovery of a couple of likely bugs.
The second is a more permanent change.

There are a couple of actual bugs (or at least worked-by-accidents) fixed by these changes.