Fixes #311 - "Debug in pdb..." button even if frozen
If the environment variable CELLPROFILER_DEBUG is set (to anything
that evaluates to true, e.g., "1"), then show the "Debug in pdb..."
button in the error dialog even if CellProfiler is frozen.

The debugger won't have a terminal to talk to when CellProfiler is
started in the usual way on a Mac. Instead, the binary inside the
application bundle must be started directly from the shell:
CellProfiler2.0.app/Contents/MacOS/CellProfiler2.0