Don't use transparent tooltip window on linux
The transparent window that covers the input modules in order to show a
tooltip does not work correctly on Linux: there, the window becomes
opaque, and the OnMove events do not work the same, so the transparent
window is not placed correctly. Rather than spend time making it work,
we skip the transparent window on Linux.