Additions to setup.py + work to make build pass on travis with pytest
Disabled test_analysis and test_analysisworker because they segfaulted on
travis. Added more of the directory tree as subpackages (it ends up in
an egg when you install). Moved the logic of CellProfiler.py to
cellprofiler.main so it can be launched as an extension point. Fixed
tests that broke with pytest on travis.