Load versioned example pipeline.
Resolves #2943

Loads a version of the ExampleFly pipeline corresponding to the running
version of CellProfiler. If no matching pipeline is found, the pipeline
is run from master.

Assumes CellProfiler/examples is tagged with the a CellProfiler version
string. When releasing a new version of CellProfiler, remember to update
and tag CellProfiler/examples with the new version.