Documentation - ColorToGray
Many changes are because it was likely originally written only for RGB and now handles other types.

I need two bits of help: 
1) pls check it's correct what I've changed from splitting -> combining
2) there are three similar settings like this: 
Select *"%(YES)s"* to extract the red channel to grayscale."""

My Q is: else what? Is it correct to add "Otherwise, the red channel will be (ignored?) (separated but stored as a color image?)" (if you know the answer please do make the change for red, green, and blue.