Documentation - SaveImages
Note that the 16/32 bit comment deleted from the main intro text does not seem so important as to mention it here (it's mentioned in two settings later), but it's fine to revert if someone thinks this is crucial to state up front. Same rationale for un-bolding 16/32 comment in one of the settings. No other setting I've come across in CP is bold, so it's hard to see why it's warranted here!