Documentation - EnhanceEdges
Is it possible to move settings to be this order in the help? 
Select the input (already first)
Name the output (already second)
Select an edge-finding method (ought to be third, right now a few settings are shown that are specific to dift methods so it's a little confusing)

I did so but feel free to revert if it causes problems.