Documentation - Input modules are 3D compatible, 3
I've already done NamesAndTypes. 
I know it's a little weird to say that these Input modules support 3D because in some cases that is sort of irrelevant to the function of the module but whatever, it's better than people thinking they are NOT 3D compatible (if left as NO) or that we forgot to make a notation for the module (if we took out the table altogether).