Use `READ COMMITTED` isolation in place of coarse grained lock
In a highly concurent batch processing environment there can easily be
100s of instances of CellProfiler.  With a decently complicated
pipeline and images collected with large frame (2K x 2K+) cameras there
can also easily be several 100 objects per image.  Under these
conditions MySQL lock timeouts will occur when using the InnoDB database
engine which is the default as of MySQL 5.7.

The coarse grained lock was instituted in 26c3841 as a fix for #1022.
While that issue does not detail the specifics it can be assumed that
the deadlocks mentioned were related to gap locking:

  https://dev.mysql.com/doc/refman/5.7/en/innodb-locking.html#innodb-gap-locks

The propensity for InnoDB to be affected by deadlocking problems is well
known.  So much so that there is a page dedicated to it in the MySQL
manual:

  https://dev.mysql.com/doc/refman/5.7/en/innodb-deadlocks.html

The ExportToDatabase module really executes no interdependent SELECT
statements in each corresponding transaction.  So, solving deadlocking
issues can be done much more elegently and concurrently by selecting an
appropriate level of transaction isolation for MySQL.  These levels are
detailed here:

  https://dev.mysql.com/doc/refman/5.7/en/innodb-transaction-isolation-levels.html

The level `READ COMMITTED` has been specifically chosen due to the
lack of interdependency in CellProfiler batch processing workflows and
the fact that gap locking is *disabled* for this isolation level.