PNG compression solves the storage issue (#3545)
* PNG compression solves the storage issue

* remove timestamp and 0 padding

I noticed since the timestamp was made into an integer that the timestamp value was often the same for each object, so it did not necessarily serve to make the object name more unique. I thought the filenames were more clear without the timestamp. Also, the zero padding suggested a maximum object count of 9999, but we've seen examples of slide scans where there are many more objects than this number. I believe removing the padding makes that naming scheme more flexible to handle an unknown number of objects.

* add option to use png or tiff

* remove time tag in test file

* Make file format options constants