Issue/3598 (#3619)
* Adapts the channel_choice widget to cps.Integer

We are dealing routinely with 30+ channel images.
Thus it would be convient to easily extract single
planes to grayscale. The previous solution was elegant
and supported that many channels, but was in practice
tedious, as selecting channel idx 20 would require
the addition of 20 channels.

* Adapted tests to channel_choice type change

The channel_choice widget is now an cps.Integer
which required a slight adaptation of some tests.

* Adapts the documentation

This adapts the documentation of `Add channel` to
reflect the fact that it is now integer values.

* Adds a test to load an old pipeline from v3

A v3 pipeline was added where images were named
such that they contained the channel number they should
represent. Like this it can be verified that channel numbers
are correctly assigned.

* Bumped version number

* Only update variable names if pipeline comes from older version

Settings from the current version do not need any update.

* Fixes formating

- Added spacing around all operators I found
- Put the class variable at the start of the class (to make issues like
seen as part of #3628 more obvious)

* Upgrades also variable_revision_number