Python3 issue 3784 (#3795)
* quick fixes to get it to run on Windows 10 64bit

* limiting change to only windows

windows registry items cannot be None anyways

* Got it to open a CP 3.1.8 cpproj file but...

does not yet load images from the cpproj file

* still doesn't fix the cppproj loading but...

was breaking here on windows. Updated to not break preferencesview in the case there are many paths that equal the abspath

* Got the cpproj's file list to populate but am not able to open images

* Got the cpproj's file list to populate but am not able to open images

at the same time, my understanding of cpprojs might be incorrect. i thought cpprojs also preserve the images so that it can be transferable between OSs but that might not be the case.

* Reverting change and using new compat equality checker instead

* Fixed #3782 (#3789)

* Python3 issue 3779 (#3791)

* quick fixes to get it to run on Windows 10 64bit

* limiting change to only windows

windows registry items cannot be None anyways

* background will be redrawn correctly, tested in windows and ubuntu

* Python3 issue 3781 (#3792)

* quick fixes to get it to run on Windows 10 64bit

* limiting change to only windows

windows registry items cannot be None anyways

* cmp keyword argument no longer exists

* Python3 issue 3783 (#3794)

* quick fixes to get it to run on Windows 10 64bit

* limiting change to only windows

windows registry items cannot be None anyways

* problem was solved by helping it correctly find the wx.text_ctrl object, was finding the wx.staticText object instead; seems like getting the font using wx.SYS_ANSI_FIXED_FONT and wx.SYS_ANSI_VAR_FONT doesn't work on windows 10 64bit. Might need to revisit

* Got it to open a CP 3.1.8 cpproj file but...

does not yet load images from the cpproj file

* still doesn't fix the cppproj loading but...

was breaking here on windows. Updated to not break preferencesview in the case there are many paths that equal the abspath

* Got the cpproj's file list to populate but am not able to open images

* Reverting change and using new compat equality checker instead

* fixed loading cppipe files using cppipe version 3,4,&5

* removed dead code