Fix WX4 GUI bugs (#3856)
* Fix broken link

* Add confirm dialog to "revert to saved" button

* Closes #3322

* Fix issue #3855

* Fix Adjust modules GUI duplication when entering test mode

* Fix error on setting change

* Fix dropfiles lacking result

FLDropTarget was expecting a result and therefore complained whenever adding a file. This eems to fix it.

* Remove console spam

Disable printing of path list to the console

* Remove console spam

Disable printing of path list to the console

* I am not fast

* Fix zoom cursor log spam on Windows

* Fix zoom/pan cursor errors

Fixes log spam when user drags outside image viewing pane.

* Fix NamesAndTypes help message

* Fix module help button

* Fix ImageMath subplots display

Error displaying subplots from the ImageMath module. Number of plots to show was producing a float.

* Update cellprofiler/gui/figure.py

Co-Authored-By: Matthew Bowden <bowdenm@spu.edu>

* Update cellprofiler/gui/figure.py

Co-Authored-By: Matthew Bowden <bowdenm@spu.edu>

* Implement AetherUnbound's suggestion

* Return bool for OnDropText