Fix table-grid binding
CornerButtonGrid is no longer a subclass of CornerButtonMixin, but
instead invokes CornerLabelRenderer, so the resulting grid object
doesn't have `fn_clicked`, `tooltip` and `label` as attributes.
Create getter/setters so it setting these properties on the grid, sets
them on the `CornerLabelRenderer` object that the grid object has as an
attribute.