Working example of library working with on/off advanced setting
I've tried to set some reasonable defaults for a non-advanced watershed, inspired by the previous implementation (using mahotas). However, segmentation is not 1:1 as a footprint value is not provided to mahotas.label. Adding a footprint here can improve the watershed output (as tested on the cells 3d image from tutorials)