Add Watershed to CellProfiler Library (#4665)
* beginning to add watershed to library

* initial complete draft

* add library module to the gui module

* making the library.module

* add watershed line

* fix typos

* change exclude border to type bool

* Add use_advanced conditional to fix tests

* Fixing watershed masking. temp

* Adding new watershed function

* intensity method requires the intensity image

* working with GUI. Need a better check for if an image has been thresholded or not

* Working example of library working with on/off advanced setting

I've tried to set some reasonable defaults for a non-advanced watershed, inspired by the previous implementation (using mahotas). However, segmentation is not 1:1 as a footprint value is not provided to mahotas.label. Adding a footprint here can improve the watershed output (as tested on the cells 3d image from tutorials)

* added initial value and settings change check

* add optional seed plot and change display to show object labels rather than RGB values

* move seed dilation to just after maxima detection and add an option to return the seed array

* add markers_mask, perform formatting, and better intensity image settings display

* add none declump option as a str instance

* make watershed mask generally available and add a no declump option

* seed generation is not available if a user provides their own markers

* initial redraft of tests

* format tests

* resize all input images, dont filter edges by default and dilate provided markers

* rename watershed_method to method and local_maxima_method to maxima_method

* use image processing code directly in tests

* streamline tests

streamline tests by not testing all possible fixture combinations
for all tests. Though, all combinations will be tested for a
subset of tests.

* only calculate distance transform if required

* add mask test

* black formatting

* remove mask test due to shared workspace between tests

* make sure non-advanced settings are passed to library when non-adv is selected

* better maxima display

* non-advanced settings test

* black formatting and undoing non-watershed formatting

* better binary image check

* re-enable seeds from an intensity image

* shared default_settings

* more sensible logic for seed generation

* revert regional seed labelling to be the same as in the original module

* set clear border False by default

* make labelling consistent between maxima methods

* maxima exclude_border=False

* catch a no-declump early, rename maxima footprint and only find maxima if not markers method

* local maxima becomes the default method

* Remove None declump_method

Originally, the None declump method would pass the thresholded image to
the skimage watershed function, which led to a different segmentation
than if the distance transform or intensity image were passed.

However, technically that's not really `None` declumping and if it was,
it's behaving differently from IDPrimary, which just returns a labelled
array.

This comit reverts just labelled the thresholded image because it begins
to get strange when you consider a marker-based watershed. In the GUI,
you can provide a marker image, but if you select None declumping it
is entirely ignored.

TLDR: Watershed should be for watershed, not only scipy.ndimage.label

* update maxima methods in test and remove none declump tests

* align parameter names

* Remove intensity-image based maxima finding in watershed

Finding maxima in an intensity image was originally added in this
refactor, but it doesn't make sense alongside the `markers` method.

Additionally, having an intensity-based maxima method uses the same
intensity image as intensity declumping, which also doesn't make sense.

A user should instead calculate their own markers from an intensity image
using the FindMaxima module, which can then be passed in to watershed
as markers.

* update watershed documentation

* blurring distance transform should only be available if distance methods are used

* add intensity-based maxima determination + tests

* add none declumping (ie. just label the thresholded image)

* better min_intensity docs based on those written in FindMaxima

* Clean unused imports

* Post-rebase fix: bring back library imports

---------

Co-authored-by: Nodar Gogoberidze <gnodar01@gmail.com>