Remove inflect as dep
we were using the inflect library in a single module, ImageMath, for a
single task, to turn a number into an ordinal word; however the inflect
library has pydantic as a dependency, which is a huge amount of baggage
for such a simple task. replace it with a simple function.